Generalized Bound to work with discrete variables
